logout

+++ show-article.php
@@ -1,5 +1,8 @@
 <?php
 require __DIR__ . '/database/database.php';
+require __DIR__ . '/database/security.php';
+$currentUser = isLoggedIn();
+
 $articleDB = require __DIR__ . '/database/models/ArticleDB.php';
 
 $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
@@ -36,10 +39,12 @@ $article = $articleDB->fetchOne($id);
                 <h1 class="article-title"><?= $article['title'] ?></h1>
                 <div class="separator"></div>
                 <p class="article-content"><?= $article['content'] ?></p>
-                <div class="action">
-                    <a class="btn btn-primary" href="./delete-article.php?id=<?= $article['id'] ?>">Supprimer</a>
-                    <a class="btn btn-primary" href="./form-article.php?id=<?= $article['id'] ?>">Editer</a>
-                </div>
+                <?php if ($currentUser && $currentUser['id'] === $article['author']) : ?>
+                    <div class="action">
+                        <a class="btn btn-primary" href="./delete-article.php?id=<?= $article['id'] ?>">Supprimer</a>
+                        <a class="btn btn-primary" href="./form-article.php?id=<?= $article['id'] ?>">Editer</a>
+                    </div>
+                <?php endif ?>
             </div>
         </div>
         <?php
